+++ src/Message.h
@@ -8,7 +8,7 @@ public:
 	Message(const unsigned int senderId = 0, const unsigned long long timestamp = 0, const std::string&& content = "");
 
 	unsigned int m_senderId;
-	unsigned long long m_timestamp;
+	long m_timestamp;
 	std::string m_content;
 };
